App.jsをコンポーネントに分割
