+++ src/components/App.js
@@ -1,74 +1,21 @@
-import React, { useState, useReducer } from "react";
+import React, { useReducer } from "react";
 
 import "bootstrap/dist/css/bootstrap.min.css";
 
-import Event from "./Event";
+import EventForm from "./EventForm";
+import Events from "./Events";
+
 import reducer from "../reducers";
 
 const App = () => {
   const [state, dispatch] = useReducer(reducer, []);
 
-  const [title, setTitle] = useState("");
-  const [body, setBody] = useState("");
-
-  const addEvent = (e) => {
-    e.preventDefault();
-    dispatch({
-      type: "CREATE_EVENT",
-      title,
-      body,
-    });
-    setTitle("");
-    setBody("");
-  };
-
-  const deleteAllEvents = (e) => {
-    e.preventDefault();
-    const result = window.confirm("全てのイベントを本当に削除してよろしいですか？");
-    if (result) {
-      dispatch({ type: "DELETE_ALL_EVENTS" });
-    }
-  };
-
-  const unCreatable = title === "" || body === "";
-
   return (
     <>
       <div className="container">
-        <h4>イベント作成フォーム</h4>
-        <form>
-          <div className="form-group">
-            <label htmlFor="fromEventTitle">タイトル</label>
-            <input id="fromEventTitle" className="form-control" value={title} onChange={(e) => setTitle(e.target.value)} />
-          </div>
-          <div className="form-group">
-            <label htmlFor="fromEventBody">ボディ</label>
-            <textarea id="fromEventBody" className="form-control" value={body} onChange={(e) => setBody(e.target.value)} />
-          </div>
-          <button className="btn btn-primary" onClick={addEvent} disabled={unCreatable}>
-            イベントを作成する
-          </button>
-          <button className="btn btn-danger" onClick={deleteAllEvents} disabled={state.length === 0}>
-            全てのイベントを削除する
-          </button>
-        </form>
+        <EventForm state={state} dispatch={dispatch} />
         <hr className="" />
-        <h4>イベント一覧</h4>
-        <table className="table table-hover">
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>タイトル</th>
-              <th>ボディ</th>
-              <th></th>
-            </tr>
-          </thead>
-          <tbody>
-            {state.map((event, index) => (
-              <Event key={index} event={event} dispatch={dispatch} />
-            ))}
-          </tbody>
-        </table>
+        <Events state={state} dispatch={dispatch} />
       </div>
     </>
   );
